chore: Remove usage of useModularDrawerVisibility in LWM
