+++ .changeset/metal-numbers-trade.md
@@ -1,0 +1,5 @@
+---
+"live-mobile": minor
+---
+
+Remove usage of useModularDrawerVisibility in LWM
