+++ apps/ledger-live-mobile/src/components/Stake/index.tsx
@@ -2,9 +2,7 @@
 import { useLayoutEffect, useRef } from "react";
 import { NativeStackNavigationProp } from "@react-navigation/native-stack";
 import { useNavigation } from "@react-navigation/native";
-import { findCryptoCurrencyById } from "@ledgerhq/cryptoassets/currencies";
-import { getCryptoAssetsStore } from "@ledgerhq/cryptoassets/state";
-import { NavigatorName, ScreenName } from "~/const";
+import { ScreenName } from "~/const";
 import type { StackNavigatorProps, BaseComposite } from "../RootNavigator/types/helpers";
 import type { StakeNavigatorParamList } from "../RootNavigator/types/StakeNavigator";
 import { useStakingDrawer } from "./useStakingDrawer";
@@ -29,7 +27,7 @@ const StakeFlow = ({ route }: Props) => {
     entryPoint: route?.params?.entryPoint,
   });
 
-  const { handleOpenStakeDrawer, isModularDrawerEnabled } = useOpenStakeDrawer({
+  const { handleOpenStakeDrawer } = useOpenStakeDrawer({
     currencies,
     sourceScreenName: "stake_flow",
   });
@@ -37,38 +35,7 @@ const StakeFlow = ({ route }: Props) => {
   const requestAccountRef = useRef(() => {});
 
   requestAccountRef.current = () => {
-    if (isModularDrawerEnabled) {
-      handleOpenStakeDrawer();
-      return;
-    } else {
-      // Fallback to traditional navigation
-      if (currencies.length === 1) {
-        Promise.resolve(
-          findCryptoCurrencyById(currencies[0]) ||
-            getCryptoAssetsStore().findTokenById(currencies[0]),
-        ).then(currency => {
-          if (!currency) return;
-          // Navigate to the second screen when there is only one currency
-          navigation.replace(NavigatorName.RequestAccount, {
-            screen: ScreenName.RequestAccountsSelectAccount,
-            params: {
-              currency,
-              onSuccess: goToAccountStakeFlow,
-              allowAddAccount: true, // if no account, need to be able to add one to get funds.
-            },
-          });
-        });
-      } else {
-        navigation.replace(NavigatorName.RequestAccount, {
-          screen: ScreenName.RequestAccountsSelectCrypto,
-          params: {
-            currencyIds: currencies,
-            allowAddAccount: true,
-            onSuccess: goToAccountStakeFlow,
-          },
-        });
-      }
-    }
+    handleOpenStakeDrawer();
   };
 
   useLayoutEffect(() => {
