+++ apps/ledger-live-mobile/src/components/Web3AppWebview/NoAccountScreen.tsx
@@ -6,41 +6,16 @@ import { AppManifest } from "@ledgerhq/live-common/wallet-api/types";
 import { CurrentAccountHistDB } from "@ledgerhq/live-common/wallet-api/react";
 import Button from "../Button";
 import { useSelectAccount } from "./helpers";
-import { OpenDrawer } from "LLM/features/ModularDrawer/types";
-import { currentRouteNameRef } from "~/analytics/screenRefs";
 type NoAccountScreenProps = {
   manifest: AppManifest;
   currentAccountHistDb?: CurrentAccountHistDB;
-  openModularDrawer?: OpenDrawer;
 };
 
-export function NoAccountScreen({
-  manifest,
-  currentAccountHistDb,
-  openModularDrawer,
-}: NoAccountScreenProps) {
-  const { onSelectAccount, onSelectAccountSuccess, currencyIds } = useSelectAccount({
+export function NoAccountScreen({ manifest, currentAccountHistDb }: NoAccountScreenProps) {
+  const { handleAddAccountPress } = useSelectAccount({
     manifest,
     currentAccountHistDb,
   });
-
-  const handleAddAccountPress = () => {
-    if (openModularDrawer) {
-      openModularDrawer({
-        currencies: currencyIds,
-        areCurrenciesFiltered: true,
-        enableAccountSelection: true,
-        onAccountSelected: onSelectAccountSuccess,
-        flow: manifest.name,
-        source:
-          currentRouteNameRef.current === "Platform Catalog"
-            ? "Discover"
-            : currentRouteNameRef.current ?? "Unknown",
-      });
-    } else {
-      onSelectAccount();
-    }
-  };
 
   return (
     <View style={styles.container}>
