+++ apps/ledger-live-mobile/src/components/Web3AppWebview/WalletAPIWebview.tsx
@@ -35,8 +35,6 @@ export const WalletAPIWebview = forwardRef<WebviewAPI, WebviewProps>(
       webviewRef,
       webviewCacheOptions,
       noAccounts,
-      isModularDrawerVisible,
-      openModularDrawer,
     } = useWebView(
       {
         manifest,
@@ -60,13 +58,7 @@ export const WalletAPIWebview = forwardRef<WebviewAPI, WebviewProps>(
       internalAppIds.includes(manifest.id) || manifest.id === WC_ID;
 
     if (!!manifest.dapp && noAccounts) {
-      return (
-        <NoAccountScreen
-          manifest={manifest}
-          currentAccountHistDb={currentAccountHistDb}
-          openModularDrawer={isModularDrawerVisible ? openModularDrawer : undefined}
-        />
-      );
+      return <NoAccountScreen manifest={manifest} currentAccountHistDb={currentAccountHistDb} />;
     }
 
     return (
