+++ apps/ledger-live-mobile/src/newArch/features/ModularDrawer/index.ts
@@ -2,7 +2,6 @@ import { ModularDrawer } from "LLM/features/ModularDrawer/ModularDrawer";
 import { ModularDrawerStep } from "LLM/features/ModularDrawer/types";
 import { useModularDrawerController } from "LLM/features/ModularDrawer/hooks/useModularDrawerController";
 import { ModularDrawerWrapper } from "LLM/features/ModularDrawer/ModularDrawerWrapper";
-import { useModularDrawerVisibility } from "@ledgerhq/live-common/modularDrawer/useModularDrawerVisibility";
 import { ModularDrawerLocation } from "@ledgerhq/live-common/modularDrawer/enums";
 
 export {
@@ -10,6 +9,5 @@ export {
   ModularDrawerStep,
   useModularDrawerController,
   ModularDrawerWrapper,
-  useModularDrawerVisibility,
   ModularDrawerLocation,
 };
